+++ core/config/routes.php
@@ -31,6 +31,9 @@ $route['^go-([a-zA-Z_-]+)$'] = "website/change_theme/$1";
 //Change language route
 $route['^change-language/([a-z]+)$'] = "website/change_language/$1";
 
+//Minified resources route
+$route['^attach/cache/resources-([a-z]{1,3})/(.+)'] = "website/minify";
+
 //The route that generates images with presets
 $route['^attach/cache/([A-Za-z0-9_]+)/([A-Za-z0-9_]+)/([0-9]+)/([a-z0-9_-]+)/([A-Za-z0-9_-]+)\.([A-z]{3,4})'] = "website/image_router/$1/$2/$3/$4/$5/$6";