+++ core/config/website.php
@@ -64,7 +64,7 @@ $config['prepend_uri_language'] = TRUE;
 /*
  * Framework version
  */
-define('BANCHA_VERSION', '0.9.8');
+define('BANCHA_VERSION', '0.9.11');
 
 /*
  * Framework name
@@ -90,11 +90,24 @@ $config['installed_themes'] = array(
 );
 
 /*
-* CACHE - CATEGORIES AND HIERARCHIES
+* CACHE - CATEGORIES, HIERARCHIES, CSS+JS MINIFIED FILES
 * Defines if the categories and hierarchies queries can be cached on the filesystem.
+* Defines also if the CSS + JS files can be cached on disk by the minifier script.
 * Feel free to disable this feature if your disk makes a lot of file read/write operations.
 */
-define('CACHE', TRUE);
+if (!defined('CACHE'))
+{
+	define('CACHE', TRUE);
+}
+
+/*
+* API ENABLED
+* Enable/Disable the API system
+*/
+if (!defined('API_ENABLED'))
+{
+	define('API_ENABLED', TRUE);
+}
 
 /*
 * SHARED API TOKEN
@@ -112,7 +125,10 @@ $config['records_per_page'] = 15;
 /* FRAMEWORK PATH
  * The library path where Bancha classes are stored
  */
-define('FRPATH', APPPATH . 'libraries' . $sep . FRNAME . $sep);
+if (!defined('FRPATH'))
+{
+	define('FRPATH', APPPATH . 'libraries' . $sep . FRNAME . $sep);
+}
 
 /*
 * XML FOLDER