+++ core/controllers/admin/api.php
@@ -17,6 +17,12 @@ Class Core_Api extends Bancha_Controller
 	public function __construct()
 	{
 	    parent::__construct();
+	    $this->load->database();
+
+	    if (API_ENABLED == FALSE)
+	    {
+	    	show_error('API are disabled.', 400);
+	    }
 
 		$this->content->set_stage($this->input->post('production') == TRUE ? FALSE : TRUE);
 	    $this->view->base = 'admin/';