+++ core/controllers/admin/contents.php
@@ -25,6 +25,7 @@ Class Core_Contents extends Bancha_Controller
   	public function __construct()
   	{
 	    parent::__construct();
+        $this->load->database();
 
 	    //We are always in staging here
 	    $this->content->set_stage(TRUE);