+++ core/controllers/admin/settings.php
@@ -20,6 +20,7 @@ Class Core_Settings extends Bancha_Controller
 	public function __construct()
 	{
 	    parent::__construct();
+	    $this->load->database();
 
 	    $this->content->set_stage(TRUE);