+++ core/controllers/admin/themes.php
@@ -19,6 +19,7 @@ Class Core_Themes extends Bancha_Controller
 	public function __construct()
 	{
 	    parent::__construct();
+	    $this->load->database();
 
 	    $this->content->set_stage(TRUE);
 	    $this->view->base = 'admin/';