+++ core/helpers/website_helper.php
@@ -2,7 +2,9 @@
 /**
  * Website Helper
  *
- * Some utilities for the website (back and front end)
+ * Some utilities for the website
+ * ------------------------------
+ * Please do not change the functions below. Instead, feel free to copy and rename them.
  *
  * @package		Bancha
  * @author		Nicholas Valbusa - info@squallstar.it - @squallstar
@@ -11,6 +13,15 @@
  * @link		http://squallstar.it
  *
  */
+
+/**
+ * Returns the Bancha singleton super-object
+ * @return CI_Controller
+ */
+function &bancha()
+{
+	return CI_Controller::get_instance();
+}
 
 /**
  * Dumps an object (or a variable)
@@ -81,7 +92,8 @@ function attach_url($str='')
  * @param $preset preset name
  * @param $append_siteurl whether to prepend or not the website url
  */
-function preset_url($path, $preset, $append_siteurl = TRUE) {
+function preset_url($path, $preset, $append_siteurl = TRUE)
+{
 	if ($path && $preset)
 	{
 		//Prototype: attach/cache/type/field/id/preset/name.ext
@@ -91,4 +103,82 @@ function preset_url($path, $preset, $append_siteurl = TRUE) {
 		return $append_siteurl ? site_url($path, FALSE) : $path;
 	}
 	return '';
+}
+
+/**
+ * Returns the path of a minified file with the provided resources
+ * @param array|string $files One or more resources
+ * @param int $version
+ * @return string
+ */
+function minify($files = array(), $version='')
+{
+	if (is_string($files))
+	{
+		$files = array($files);
+	}
+	$current_theme = get_instance()->view->theme;
+
+	$ext = explode('.', $files[0]);
+	$ext = $ext[count($ext)-1];
+
+	$folder = config_item('attach_out_folder') . 'cache/resources-'.$ext.'/' . md5($current_theme.implode(',', $files));
+
+	$params = '?t=' . $current_theme . '&f=' . urlencode(implode(',', $files));
+	if ($version != '')
+	{
+		$folder.= '.' . $version;
+		$params.= '&v=' . $version;
+	}
+
+	$folder.= '.' . $ext;
+
+	return site_url($folder . $params, FALSE);
+}
+
+/**
+ * This function tries to generate the detail url of the given Record
+ * @param Record|string $object A record object, or just the slug to append
+ */
+function semantic_url($object = '')
+{
+	$CI = & get_instance();
+	$view = & $CI->view;
+	$record = $view->get('record');
+	$page = $view->get('page');
+
+	$current_uri = $CI->uri->uri_string;
+
+	if ($object instanceof Record)
+	{
+		$uri = $object->get('uri');
+	} else {
+		$uri = (string)$object;
+	}
+
+	if ($record)
+	{
+		if ($page->get('action') == 'single')
+		{
+			//We are in the detail page of a record. Let's replace the current detail URI with the new one
+			$parent_uri = rtrim($current_uri, $record->get('uri'));
+			return site_url($parent_uri . $uri);
+		}
+	} else {
+		//We are in a page
+		if ($page->get('action') == 'list')
+		{
+			//If the page is listing some records, we just add the new URI to the current one
+			return site_url(rtrim($current_uri, '/') . '/' . $uri);
+		}
+	}
+
+	//If we are here, is a normal page. We have to try to find another page that is listing this record
+	if ($page && $object instanceof Record)
+	{
+		return site_url($CI->pages->get_semantic_url($object->_tipo) . '/' . $uri);
+	} else {
+		//No more things to try on
+		return '#';
+	}
 }