+++ themes/minimal/views/templates/home.php
@@ -13,6 +13,7 @@
 if ( ! defined('BASEPATH')) exit('No direct script access allowed');
 
 $this->view->render('header');
+$posts = $this->records->type('Blog')->get();
 
 ?>
 
@@ -22,6 +23,15 @@ $this->view->render('header');
 		<div class="body">
 		Nunc eget nunc libero. Nunc commodo euismod massa quis vestibulum. Proin mi nibh, dignissim a pellentesque at, ultricies sit amet sapien. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Quisque vel lorem eu libero laoreet facilisis. Aenean placerat, ligula quis placerat iaculis, mi magna luctus nibh, adipiscing pretium erat neque vitae augue. Quisque consectetur odio ut sem semper commodo. Maecenas iaculis leo a ligula euismod condimentum. Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Ut enim risus, rhoncus sit amet ultricies vel, aliquet ut dolor. Duis iaculis urna vel massa ultricies suscipit. Phasellus diam sapien, fermentum a eleifend non, luctus non augue. Quisque scelerisque purus quis eros sollicitudin gravida. Aliquam erat volutpat. Donec a sem consequat tortor posuere dignissim sit amet at ipsum.<br /><br />
 	
+
+		<h2>Random posts</h2>
+
+		<?php foreach ($posts as $post) {
+			echo '<a href="' . semantic_url($post) . '">' . $post->get('title') . '</a>' . br(2);
+		} ?>
+
+		<hr />
+
 		<?php
 	
 		$sharer = $this->load->module('sharer');